ESC-532 introduce Connector trait and apply to all connectors
